feat: fix all bugs
fix s21_pow funtional

change style in all modules

delete unused defines

use int*_t types

format all modules with clang

correct tasks.json to debug main.c

+++ src/modules/s21_abs.c
@@ -18,7 +18,7 @@
  * @return int - result of calculation.
  */
 int s21_abs(int x) {
-  int shift = GET_SHIFT(x);
+  uint8_t shift = GET_SHIFT(x);
 
   if (x >> shift) {
     x = ~x + 1;
